requirements.txt
click==8.0
scipy==1.11.1
tqdm==4.62.2
ninja==1.10.2
matplotlib==3.5.3
imageio==2.9.0
numpy==1.22.4
pytorch_lightning==2.0.0
timm==0.6.12
glfw==2.2.0
pyopengl==3.1.5
imageio-ffmpeg==0.4.3
pyspng
psutil
mrcfile
tensorboard
transformers==4.26.1
open-clip-torch==2.16.0
einops==0.6.0
opencv-python
scikit-image
tensorflow==2.10.0
mtcnn
dlib==19.24.0
moviepy==1.0.3
gradio== 3.30.0